# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Drafenex backend API",
        "title": "Drafenex API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/organizer/register": {
            "post": {
                "description": "Create organizer onboarding profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizer"
                ],
                "summary": "Create organizer",
                "parameters": [
                    {
                        "description": "Organizer input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrganizerInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateOrganizerInput": {
            "type": "object",
            "properties": {
                "addressLine1": {
                    "type": "string"
                },
                "addressLine2": {
                    "type": "string"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        }
    }
}